Add game() function
The game() function executes the playRound function 5 times in a row.
That is now we can play five times without interruption with the computer.
At the end we use the finalScore() function to display the results.

N.B Add some problems using variables, especially how and when to use the
let statement.For now I just used "variable = value" to declare a variable
as it's not possible to change the value of a variable after declaring it with let.
Really to pay attention about this one.