Add event listener to every button and modifications to game()
Add event listener to every button so that we can start play a round
every we click on one of them.
To accomplish this we have used the game() function as a callback one.
So, it was necessary to make some modifications to the game() functions.
The most important one being using the event object as argument so that
we can read the user choice.